backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyMuPDF==1.23.5
python-dotenv==1.0.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-anthropic==0.1.1
langchain-community==0.0.10